backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-dotenv
pydantic[email]
python-jose[cryptography]
passlib
bcrypt==3.2.2
python-multipart
celery
redis
google-generativeai
twilio
requests
